LayoutTests/imported/w3c/web-platform-tests/css/css-contain/contain-layout-dynamic-001-expected.txt


PASS contain: none
PASS contain: layout
FAIL switching contain from none to layout assert_equals: vertical baseline suppressed expected true but got false
FAIL switching contain from layout to none assert_equals: vertical baseline suppressed expected false but got true
X
X
X
X
